adk-agents/requirements.txt
# Google Agent Development Kit
google-adk[a2a]>=0.1.0

# Core dependencies
google-generativeai>=0.3.0
google-cloud-firestore>=2.13.0
google-cloud-functions>=1.15.0

# HTTP and async
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.1.0
tenacity>=8.2.0
